+++ app/Http/Controllers/Library/LendingController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Library;
 
+use App;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Library\StoreExtendBook;
 use App\Http\Requests\Library\StoreExtendBookToPerson;
@@ -14,6 +15,7 @@ use App\Models\Library\LibraryLending;
 use App\Models\People\Person;
 use App\Settings\Library\DefaultLendingDurationDays;
 use Carbon\Carbon;
+use Languages;
 use Setting;
 
 class LendingController extends Controller
@@ -76,6 +78,7 @@ class LendingController extends Controller
             'person' => $person,
             'lendings' => $person->bookLendings()->whereNull('returned_date')->orderBy('return_date', 'asc')->get(),
             'default_extend_duration' => Setting::get('library.default_lending_duration_days', DefaultLendingDurationDays::DEFAULT_VALUE),
+            'languages' => Languages::lookup(null, App::getLocale()),
         ]);
     }
 
@@ -129,7 +132,7 @@ class LendingController extends Controller
             $book->title = $request->title;
             $book->author = $request->author;
             $book->isbn = $request->isbn;
-            $book->language = $request->language;
+            $book->language_code = $request->language_code;
             $book->save();
         } else {
             $book = LibraryBook::findOrFail($request->book_id);