+++ app/Navigation/ContextButtons/Library/LibraryReportContextButtons.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Navigation\ContextButtons\Library;
+
+use App\Navigation\ContextButtons\ContextButtons;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\View\View;
+
+class LibraryReportContextButtons implements ContextButtons
+{
+    public function getItems(View $view): array
+    {
+        return [
+            'back' => [
+                'url' => route('library.lending.index'),
+                'caption' => __('app.close'),
+                'icon' => 'times-circle',
+                'authorized' => Gate::allows('operate-library'),
+            ],
+        ];
+    }
+
+}