+++ database/factories/LibraryModelFactory.php
@@ -8,12 +8,11 @@ use App\Models\Library\LibraryBook;
  * @var \Illuminate\Database\Eloquent\Factory $factory
  */
 $factory->define(LibraryBook::class, function (Faker\Generator $faker) {
-    $lc = $faker->optional(0.8)->languageCode;
-    $language = $lc != null ? collect(\Languages::lookup([$lc]))->first() : null;
+    $sentence = $faker->sentence;
     return [
-        'title' => $faker->sentence,
+        'title' => substr($sentence, 0, strlen($sentence) - 1),
         'author' => $faker->optional(0.9)->name,
-        'language' => $language,
+        'language_code' => $faker->optional(0.8)->languageCode,
         'isbn' => $faker->optional(0.6)->isbn10,
     ];
 });